test: fold eight Nav classes into two, and cover what they missed
Eight files, nineteen tests, and eight copies of the same twelve-line
setUp. Now two files, the same nineteen tests, and 29 more assertions.

Two of the eight were not testing anything. NavNullLinkTest and
NavFontResetTest installed their template as the $template global, but
Template::templateReplace() reads the Template singleton, so buildNavs()
returned an empty string for both. NavFontResetTest's every assertion is
assertStringNotContainsString, which an empty string satisfies; and
NavNullLinkTest compared two calls that both rendered nothing. Their
cases are kept against a real template, each now paired with a positive
control so a silent empty render cannot pass for a correct one again.

NavSortTest had a genuine gap behind a misleading name. The three
preferences are named one level off from what they control --
sortedmenus orders the headers, navsort_headers the sub-headers,
navsort_subheaders the items -- and testHeaderAscendingSorting set
navsort_headers while leaving sortedmenus unset. Headers were therefore
never sorted, and its assertion held only because insertion order
happened to match.

The resulting hole is narrow, so it is worth saying exactly what it was:
breaking ascending sorting outright was caught by testAscendingSorting,
and removing header sorting outright by testHeaderDescendingSorting.
What nothing caught was ascending sorting failing at the header level
alone. Forcing $sectionOrder from 'asc' to 'off' in buildNavs() leaves
the old file green and fails the new testHeadersSortAscending.

Renaming that test truthfully also turned up a second overstatement:
testSortedMenusPreferenceDisablesSorting did not disable sorting, only
the header sort -- the items go on sorting under their own preference,
which is what its own assertions said. Both now carry names that match.

Dropped: the Reflection method in NavAccessKeyHolidayRegressionTest,
which reached into two private static helpers and blocked any refactor of
them. Its two siblings exercise the same behaviour through
privateAddNav() and stay. Also dropped an assertStringContainsString
('</span>') that any template in the file satisfies.

Default, reverse and two random orders green at 5364 assertions (5335 on
master); phpstan clean, phpcs unchanged.
